requirements.txt
flask==1.0.3
firebase-admin==2.17.0
PyJWT==1.7.1
opencensus==0.6.0
opencensus-ext-flask==0.3.1
opencensus-ext-stackdriver==0.4.0
opencensus-ext-google-cloud-clientlibs==0.1.2
